+++ library/feature/faq/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/faq/data/local/FaqLocalDataSource.kt
@@ -18,8 +18,8 @@
 package com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.data.local
 
 import android.content.Context
-import com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.domain.models.FaqId
-import com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.domain.models.FaqItem
+import com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.data.models.FaqId
+import com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.data.models.FaqItem
 import com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.R
 
 /**