+++ library/feature/faq/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/faq/data/models/FaqItem.kt
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.domain.models
+package com.mihaicristiancondrea.android.libs.apptoolkit.feature.faq.data.models
 
 @JvmInline
 value class FaqId(val value: String)